Revised routing structure and renamed user state
This commit implements a more efficient routing system with Suspense and BrowserRouter, as well as the introduction of a new ProtectedRoute component. In addition, the 'user' slice of the Redux state has been renamed as 'currentUser' to more accurately reflect its function within the application. These changes aim to improve security, user experience, and code understandability.