Add new game handler and enhance modal window
Introduced a new component `HandleNewGame` and performed changes to the `ModalWindow` for improved user interaction. `HandleNewGame` allows users to create a new game, while changes to the `ModalWindow` include the application of backdrop-filter for visual improvement and a new hook `useOutsideClicker` that provides functionality to close the modal by clicking outside of it. Also, modifications were made to the `Header` component to include the newly created `HandleNewGame` component.