Add Checkbox feature and modify form designs
In this update, a new Checkbox component has been added and used in several forms. CSS changes have also been instituted, such as modifying the class name of the SignInForm and adding new styling for the Checkbox component. Moreover, an option for private game rooms has been introduced in the CreateRoomForm.