Add InviteUserToGame component and update ModalWindow
This commit introduces a new component, InviteUserToGame, with functionality for inviting users to a game. It also simplifies the ModalWindow component by removing unnecessary React Fragments. Additionally, it includes the usage of InviteUserToGame in the PlayerStatePanel component and a new graphql query for searching users. These changes improve user interaction and provide better organization and readability in the code.