Add SearchDropDown component to InviteUserToGame modal
A new SearchDropDown component has been introduced and added into the InviteUserToGame modal window. This component handles search operations and displays the search results in a dropdown list. In addition, the associated CSS styling for the new search dropdown has been created. The component will be triggered when users need to invite other users to a game, making the invitation process more interactive and user-friendly.