Refactor InviteUserToGame and form components
Replaced SearchDropDown with the InviteUsersForm in the InviteUserToGame enabling users to search for and invite other players. Adjusted SignInForm and SignUpForm components to use a local FormValues type for better flexibility. Also created a new InviteUsersForm component to handle invitation operations. Changes were made to the SearchDropDown and types file as well. This update enhances the inviting user functionality and simplifies forms implementation.