Implement invite players mutation in game
This commit introduces the `invitePlayers` mutation which is used to handle inviting multiple users to join a game. Changes are made in `InviteUsersForm.tsx`, `InviteUserToGame.tsx`, and `game.ts`, allowing users to be selected and an invite to be triggered by passing user ids and the game id to the mutation.