Refactor invitePlayers method for better data management
The invitePlayers method in the InviteUsersForm component has been updated to take an object with game details instead of just the id. This helps improve data handling by providing all necessary game-related info in one object. This refactoring includes changes in the GraphQL mutation to align with the updated input format.