Refactor game and user states into separate reducers and selectors
This commit encompasses several changes: the major one is the splitting of the `game` and `user` states into separate reducer and selector files to improve organization and readability. It also includes a new chat reducer added to the store. Lastly, various components were adjusted to use the new selectors for the `game` state instead of directly accessing state properties. As a result, our redux store is now more modular, and the code is easier to maintain and understand.