Refactor game state panel and integrate chat messages
Modified part of the GameStatePanel module to include 'CurrentGame' type instead of the 'Game' type from the game selector. Additionally, new modules and functionality were added to incorporate the display of chat messages on the GameChatPanel. The new 'ChatMessage' molecule and 'selectMessages' selector were created and used in the GameChatPanel to enable this addition.